adding changes

+++ src/main/java/com/studentsmanagement/service/StudentService.java
@@ -1,12 +1,12 @@
 package com.studentsmanagement.service;
 
+import com.studentsmanagement.exception.StudentNotFoundException;
 import com.studentsmanagement.model.Student;
 import com.studentsmanagement.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class StudentService {
@@ -20,8 +20,9 @@ public class StudentService {
         return studentRepository.findAll();
     }
 
-    public Optional<Student> getStudentByid(String id){
-        return studentRepository.findById(id);
+    public Student getStudentByid(String id){
+        return studentRepository.findById(id)
+                .orElseThrow(()->new StudentNotFoundException("Student with ID"+id+"Not found."));
 
     }
 
@@ -31,29 +32,40 @@ public class StudentService {
     }
 
     public void deleteStudent(String id){
+//        studentRepository.deleteById(id);
+        if(!studentRepository.existsById(id)){
+            throw new StudentNotFoundException("Student with id "+id+"Not found for deletion.");
+        }
         studentRepository.deleteById(id);
     }
 
-    public  List<Student> getstudentbyname(String name){
-       return studentRepository.findByName(name);
-    }
+//    public  List<Student> getstudentbyname(String name){
+//       return studentRepository.findByName(name);
+//    }
 
 //    public  List<Student> getStudentbyagegreaterthan(int age){
 //            return  studentRepository.findbyagegreaterthan(age);
 //    }
-//
-//
+
 //    public List<Student> getstudentsbycity(String city){
 //        return studentRepository.findbycity(city);
 //    }
 
 
     public  Student getstudentbyemail(String email){
-        return studentRepository.findbyemail(email);
+        Student student = studentRepository.findbyemail(email);
+        if(student==null){
+            throw  new StudentNotFoundException("No Student with the email exist"+email);
+        }
+        return student;
     }
 
 
-//    public List<Student> getstudentbetweenages(int minage,int maxage){
-//        return studentRepository.findstudentbetweenage(minage,maxage);
-//    }
+    public List<Student> getstudentbetweenages(int minage,int maxage){
+        List<Student> students = studentRepository.findstudentbetweenage(minage, maxage);
+        if(students.isEmpty()){
+            throw  new StudentNotFoundException("No Student found between ages "+minage+"and "+maxage);
+        }
+        return students;
+    }
 }
